build-system-ci-logging

+++ packages/scripts/lib/utils/spinner.ts
@@ -1,21 +1,17 @@
 import yoctoSpinner from "yocto-spinner"
 
-export const spinner = yoctoSpinner({ text: "" })
+const _spinner = yoctoSpinner({ text: "" })
 
 const ogLog = console.log.bind(console)
 const ogWarn = console.warn.bind(console)
 const ogError = console.error.bind(console)
-const oTable = console.table.bind(console)
+const ogTable = console.table.bind(console)
 
-const ogStart = spinner.start.bind(spinner)
-const ogSuccess = spinner.success.bind(spinner)
-const ogStop = spinner.stop.bind(spinner)
-
-function spinnerOutputFunction(f: (...msgs: unknown[]) => void) {
+function spinnerOutputFunction(f: (typeof console)["log" | "warn" | "error" | "table"]) {
     return (...msgs: unknown[]) => {
-        spinner.clear()
+        _spinner.clear()
         f(...msgs)
-        spinner.start()
+        _spinner.start()
     }
 }
 
@@ -26,7 +22,7 @@ function installSpinnerOutputFunctions() {
     console.log = spinnerOutputFunction(ogLog)
     console.warn = spinnerOutputFunction(ogWarn)
     console.error = spinnerOutputFunction(ogError)
-    console.table = spinnerOutputFunction(oTable as (...msgs: unknown[]) => void)
+    console.table = spinnerOutputFunction(ogTable)
     started = true
 }
 
@@ -34,23 +30,40 @@ function uninstallSpinnerOutputFunctions() {
     console.log = ogLog
     console.warn = ogWarn
     console.error = ogError
+    console.table = ogTable
     started = false
 }
 
-spinner.start = (text?: string) => {
-    installSpinnerOutputFunctions()
-    ogStart(text)
-    return spinner
-}
-
-spinner.success = (text?: string) => {
-    uninstallSpinnerOutputFunctions()
-    ogSuccess(text)
-    return spinner
-}
-
-spinner.stop = (text?: string) => {
-    uninstallSpinnerOutputFunctions()
-    ogStop(text)
-    return spinner
+export const spinner = {
+    start(text?: string) {
+        if (process.env.CI) {
+            console.log(text)
+        } else {
+            installSpinnerOutputFunctions()
+            _spinner.start(text)
+        }
+    },
+    stop(text?: string) {
+        if (process.env.CI) {
+            console.log(text)
+        } else {
+            uninstallSpinnerOutputFunctions()
+            _spinner.stop(text)
+        }
+    },
+    success(text?: string) {
+        if (process.env.CI) {
+            console.log(text)
+        } else {
+            uninstallSpinnerOutputFunctions()
+            _spinner.success(text)
+        }
+    },
+    setText(text: string) {
+        if (process.env.CI) {
+            console.log(text)
+        } else {
+            _spinner.text = text
+        }
+    },
 }
